Update menu.html
I was getting an invalid static map api key error for the map. This is a new problem. I'm not sure if its because the site is on a server versus localhost, or if google has changed something.
I added a static map api for amazon server browser to my code. This worked in a test browser.